+++ modules/server-node/src/services/store.ts
@@ -1034,7 +1034,7 @@ export class PrismaStore implements IServerNodeStore {
       return undefined;
     }
 
-    // not ideal, but if the channel has been detatched we need to re-attach it separatedly... todo: use join queries
+    // not ideal, but if the channel has been detatched we need to re-attach it separatedly... todo: use join queries #430
     if (!transfer.channel) {
       const channel = await this.prisma.channel.findUnique({ where: { channelAddress: transfer.channelAddressId } });
       transfer.channel = channel;
@@ -1053,7 +1053,7 @@ export class PrismaStore implements IServerNodeStore {
       return undefined;
     }
 
-    // not ideal, but if the channel has been detatched we need to re-attach it separatedly... todo: use join queries
+    // not ideal, but if the channel has been detatched we need to re-attach it separatedly... todo: use join queries #430
     if (!transfer.channel) {
       const channel = await this.prisma.channel.findUnique({ where: { channelAddress: transfer.channelAddressId } });
       transfer.channel = channel;